- Remove .json file extension - Make CI check we download a black wheel and that it gets deleted - Attempt to cleanup legacy JSON symlink - Fix delete race when canon name == argument name

+++ test_runner.py
@@ -27,6 +27,15 @@ MIRROR_BASE = MIRROR_ROOT / "web"
 TGZ_SHA256 = "bc9430dae93f8bc53728773545cbb646a6b5327f98de31bdd6e1a2b2c6e805a9"
 TOX_EXE = Path(which("tox") or "tox")
 
+# Make Global so we can check exists before delete
+A_BLACK_WHL = (
+    MIRROR_BASE
+    / "packages"
+    / "30"
+    / "62"
+    / "cf549544a5fe990bbaeca21e9c419501b2de7a701ab0afb377bc81676600"
+    / "black-19.3b0-py36-none-any.whl"
+)
 
 def check_ci() -> int:
     black_index = MIRROR_BASE / "simple/b/black/index.html"
@@ -34,7 +43,9 @@ def check_ci() -> int:
     peerme_json = MIRROR_BASE / "json/peerme"
     peerme_tgz = (
         MIRROR_BASE
-        / "packages/8f/1a/"
+        / "packages"
+        / "8f"
+        / "1a"
         / "1aa000db9c5a799b676227e845d2b64fe725328e05e3d3b30036f50eb316"
         / "peerme-1.0.0-py36-none-any.whl"
     )
@@ -60,6 +71,10 @@ def check_ci() -> int:
         print(f"{EOP} {black_index} exists ... delete failed?")
         return 73
 
+    if A_BLACK_WHL.exists():
+        print(f"{EOP} {A_BLACK_WHL} exists ... delete failed?")
+        return 74
+
     rmtree(MIRROR_ROOT)
 
     print("Bandersnatch PyPI CI finished successfully!")
@@ -75,6 +90,11 @@ def do_ci(conf: Path) -> int:
     cmds = (str(BANDERSNATCH_EXE), "--config", str(conf), "--debug", "mirror")
     print(f"bandersnatch cmd: {' '.join(cmds)}")
     run(cmds, check=True)
+
+    print(f"Checking if {A_BLACK_WHL} exists")
+    if not A_BLACK_WHL.exists():
+        print(f"{EOP} {A_BLACK_WHL} does not exist after mirroring ...")
+        return 68
 
     print("Starting to deleting black from mirror ...")
     del_cmds = (
